# api/openapi/api.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Notebook App
  description: TODO
  contact:
    name: Darek Stopka
servers:
  - url: https://{hostname}}/api
    variables:
      hostname:
        default: localhost
security:
  - bearerAuth: []
paths:
  /users/me:
    get:
      operationId: getCurrentUser
      responses:
        "200":
          description: user successfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUser"
      responses:
        "204":
          description: user successfully updated
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /notebooks:
    get:
      operationId: getNotebooks
      parameters:
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Page"
      responses:
        "200":
          description: notebooks sucessfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebooks"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createNotebook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostNotebook"
      responses:
        "201":
          description: notebook sucessfully created
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /notebooks/{notebookUUID}:
    get:
      operationId: getNotebook
      parameters:
        - $ref: "#/components/parameters/NotebookID"
      responses:
        "200":
          description: notebook suffessfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebook"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: UpdateNotebook
      parameters:
        - $ref: "#/components/parameters/NotebookID"
      requestBody:
        description: notebook successfully updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchNotebook"
      responses:
        "204":
          description: TODO
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteNotebook
      parameters:
        - $ref: "#/components/parameters/NotebookID"
      responses:
        "204":
          description: notebook successfully deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /notebooks/{notebookUUID}/notes:
    get:
      operationId: getNotebookNotes
      parameters:
        - $ref: "#/components/parameters/NotebookID"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Page"
      responses:
        "200":
          description: notebook's notes suffessfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notes"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /notes:
    get:
      operationId: getNotes
      parameters:
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Page"
      responses:
        "200":
          description: notes successfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notes"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createNote
      requestBody:
        description: note successfully created
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostNote"
      responses:
        "201":
          description: TODO
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /notes/{noteUUID}:
    get:
      operationId: getNote
      parameters:
        - $ref: "#/components/parameters/NoteID"
      responses:
        "200":
          description: note successfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateNote
      parameters:
        - $ref: "#/components/parameters/NoteID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchNote"
      responses:
        "204":
          description: note successfully updated
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteNote
      parameters:
        - $ref: "#/components/parameters/NoteID"
      responses:
        "204":
          description: note successfully deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /images/presigned-upload-url:
    post:
      operationId: getPresignedUploadURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetPresignedUploadURL"
      responses:
        "200":
          description: presign upload url successfully generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUploadURL"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserBase:
      type: object
      properties:
        avatarURL:
          type: string
          format: uri

    User:
      description: represents a single user
      allOf:
        - $ref: "#/components/schemas/UserBase"
        - type: object
          properties:
            name:
              type: string
            role:
              type: string
          required:
            - name
            - role
            - avatarURL

    PatchUser:
      $ref: "#/components/schemas/UserBase"

    NotebookBase:
      type: object
      properties:
        title:
          type: string
        icon:
          $ref: "#/components/schemas/Icon"
        description:
          type: string

    Notebook:
      description: a represents a single notebook
      allOf:
        - $ref: "#/components/schemas/NotebookBase"
        - type: object
          properties:
            uuid:
              type: string
              format: uuid
            createdTime:
              type: string
              format: date-time
            lastEditedTime:
              type: string
              format: date-time
            notesNumber:
              type: integer
          required:
            - uuid
            - createdTime
            - lastEditedTime
            - notesNumber
            - title
            - description

    Notebooks:
      description: represents a list of notebooks
      type: object
      properties:
        notebooks:
          type: array
          items:
            $ref: "#/components/schemas/Notebook"
      required: [notebooks]

    PostNotebook:
      $ref: "#/components/schemas/NotebookBase"

    PatchNotebook:
      $ref: "#/components/schemas/NotebookBase"

    NoteBase:
      type: object
      properties:
        icon:
          $ref: "#/components/schemas/Icon"
        title:
          type: string
        content:
          type: string
        externalURL:
          type: string
          format: uri
        date:
          type: string
          format: date-time
        notebook:
          type: object
          properties:
            uuid:
              type: string
              format: uuid

    Note:
      description: represents a single note
      allOf:
        - $ref: "#/components/schemas/NoteBase"
        - type: object
          properties:
            uuid:
              type: string
              format: uuid
            createdTime:
              type: string
              format: date-time
            lastEditedTime:
              type: string
              format: date-time
            notebook:
              type: object
              properties:
                uuid:
                  type: string
                  format: uuid
                title:
                  type: string
                icon:
                  $ref: "#/components/schemas/Icon"
              required:
                - uuid
                - title
          required:
            - uuid
            - createdTime
            - lastEditedTime
            - notebook
            - title
            - content
            - externalURL

    Notes:
      description: represents a list of notes
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
      required: [notes]

    PostNote:
      $ref: "#/components/schemas/NoteBase"

    PatchNote:
      $ref: "#/components/schemas/NoteBase"

    Icon:
      description: represents a single icon model
      $ref: "#/components/schemas/Emoji"

    Emoji:
      type: object
      properties:
        emoji:
          type: string
      required: [emoji]

    GetPresignedUploadURL:
      type: object
      properties:
        contentLength:
          type: integer
        contentType:
          type: string
        name:
          type: string
      required:
        - contentLength
        - contentType
        - name

    PresignedUploadURL:
      type: object
      properties:
        URL:
          type: string
        signedGetURL:
          type: string
        signedPutURL:
          type: string
        putHeaders:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
            required:
              - name
              - value
      required:
        - URL
        - signedGetURL
        - signedPutURL

    Error:
      description: represents an error returned by the service
      type: object
      properties:
        slug:
          description: error identifier
          type: string
        message:
          description: error message
          type: string
        status:
          description: HTTP status code
          type: integer
      required:
        - slug
        - message
        - status

  parameters:
    NotebookID:
      name: notebookUUID
      in: path
      required: true
      description: represents UUID of the choosen notebook object
      schema:
        type: string
        format: uuid

    NoteID:
      name: noteUUID
      in: path
      required: true
      description: represents UUID of the choosen note object
      schema:
        type: string
        format: uuid

    Sort:
      name: sort
      in: query
      description: represents a sort query parameter
      schema:
        type: string

    PageSize:
      name: pageSize
      in: query
      description: represents a page size query parameter
      schema:
        type: integer

    Page:
      name: page
      in: query
      description: represents an page number query parameter
      schema:
        type: integer